Move facets.py to models.files submodule
