+++ cidc_api/models/__init__.py
@@ -1,2 +1,3 @@
 from .models import *
+from .files import *
 from .schemas import *
